Frontti

+++ frontend/src/pages/ownpage.jsx
@@ -1,19 +1,28 @@
 import React, { useState, useEffect } from 'react'
 import axios from 'axios'
 import axiosClient from '../axios.js'
+import '../index.css'
 
 const OwnPage = ({ isLoggedIn: propIsLoggedIn }) => {
   const [username, setUsername] = useState('')
   const [email, setEmail] = useState('')
   const [telegram, setTelegram] = useState('')
   const [role, setRole] = useState('5')
+
   const [organisations, setOrganisations] = useState([])
   const [selectedOrg, setSelectedOrg] = useState(null)
   const [orgPassword, setNewOrgPassword] = useState('')
   const [confirmOrgPassword, setConfirmOrgPassword] = useState('')
   const [isChecked, setIsChecked] = useState(false)
   const [setCheckedOrgs] = useState({})
+
+  const [organization_email, setOrganizationEmail] = useState('')
+  const [organization_name, setOrganizationName] = useState('')
+  const [organization_homepage, setOrganizationHomePage] = useState('')
+  const [organization_size, setOrganizationSize] = useState('1')
+
   const [isLoggedIn, setIsLoggedIn] = useState(propIsLoggedIn)
+  const [error, setError] = useState('')
 
   const API_URL = import.meta.env.API_URL || 'http://localhost:8000'
 
@@ -26,6 +35,7 @@ const OwnPage = ({ isLoggedIn: propIsLoggedIn }) => {
       setEmail(loggedUser.email);
       setTelegram(loggedUser.telegram);
       setRole(loggedUser.role);
+      getOrganisations()
     }
   }, [propIsLoggedIn]);
 
@@ -80,6 +90,7 @@ const OwnPage = ({ isLoggedIn: propIsLoggedIn }) => {
           onChange={(e) => setTelegram(e.target.value)}
         />
       </div>
+      <br />
       <div>
         Rooli: {role}
       </div>
@@ -90,7 +101,7 @@ const OwnPage = ({ isLoggedIn: propIsLoggedIn }) => {
         Tyyppi:
       </div>
       <div>
-        Myöntämis päivä:
+        Myöntämispäivä:
       </div>
       <br />
       <button onClick={handleUserDetails} className="create-user-button" type="button">
@@ -114,6 +125,7 @@ const OwnPage = ({ isLoggedIn: propIsLoggedIn }) => {
     setIsChecked(event.target.checked)
   }
 
+  // Handles the user info update when the 'Vahvista Muutokset' button is clicked
   const handleUserDetails = (event) => {
     event.preventDefault();
   
@@ -127,15 +139,21 @@ const OwnPage = ({ isLoggedIn: propIsLoggedIn }) => {
     const user_id = loggedUser.id;
   
     console.log(user_id);
-  
-    axiosClient.put(`/users/update/${user_id}/`, details)
+
+    const confirmUpdate = window.confirm("Oletko varma, että haluat päivittää käyttäjätietojasi?")
+
+    if (confirmUpdate) {
+      axiosClient.put(`/users/update/${user_id}/`, details)
       .then(response => {
         console.log('User details updated successfully:', response.data);
         localStorage.setItem('loggedUser', JSON.stringify(response.data))
       })
       .catch(error => {
         console.error('Error updating user details:', error);
-      });
+      })
+    } else {
+      console.log("User cancelled the update.")
+    }
   }
 
   // Shows more detailed information of the organizations
@@ -178,6 +196,10 @@ const OwnPage = ({ isLoggedIn: propIsLoggedIn }) => {
                 onChange={handleCheckboxChange} />
             </label>
           </div>
+          <br />
+          <button className="login-button" type="button">
+            Poista järjestö
+          </button>
         </div>
       )
     }
@@ -193,13 +215,80 @@ const OwnPage = ({ isLoggedIn: propIsLoggedIn }) => {
         {organisations.map(org => (
           <li key={org.id}>
             {org.name}
-            <button onClick={() => toggleDetails(org.id)}>View</button>
+            <button className= "login-button" onClick={() => toggleDetails(org.id)}>View</button>
             {renderOrganizationDetails(org.id)}
           </li>
         ))}
       </ul>
     </div>
   );
+
+  const createOrganization = () => (
+    <form>
+      <h2>Luo uusi järjestö</h2>
+        <div>
+          Järjestön nimi:
+          <input
+            type="text"
+            id="name"
+            value={organization_name}
+            onChange={(e) => setOrganizationName(e.target.value)}/>
+        </div>
+        <div>
+          Sähköposti:
+          <input
+            type="text"
+            id= "email"
+            value={organization_email}
+            onChange={(e) => setOrganizationEmail(e.target.value)}/>
+        </div>
+        <div>
+          Kotisivut:
+          <input
+          type="text"
+          id= "homepage"
+          value={organization_homepage}
+          onChange={(e) => setOrganizationHomePage(e.target.value)} />
+        </div>
+        <div>
+          Koko:
+          <input
+          type= "text"
+          id= "size"
+          value={organization_size}
+          onChange={(e) => setOrganizationSize(e.target.value)}/>
+        </div>
+        <br />
+      <button className="create-user-button" type="button" onClick={handleCreateOrganization}>
+        Luo järjestö
+      </button>
+    </form>
+  )
+
+  const handleCreateOrganization = () => (
+    axios.get(`http://localhost:8000/organizations/?email=${organization_email}`)
+      .then(response => {
+        const existingOrganizations = response.data
+        if (existingOrganizations.some(org => org.email === email)) {
+          setError('Sähköposti on jo käytössä.')
+        } else {
+          const organizationObject = { name: organization_name, email: organization_email, homepage: organization_homepage, size: organization_size }
+          console.log(organizationObject)
+          axios.post('http://localhost:8000/organizations/', organizationObject)
+            .then(response => {
+              console.log(response)
+              console.log('Organization created successfully!')
+
+            })
+            .catch(error => {
+              console.error('Error creating account:', error);
+            });
+        }
+      })
+      .catch(error => {
+        console.error('Error checking email:', error)
+      })
+  )
 
   return (
     <div>
@@ -209,6 +298,7 @@ const OwnPage = ({ isLoggedIn: propIsLoggedIn }) => {
           <div id="leftleft_content">
             {userPage()}
             {organisationPage()}
+            {role === 1 && createOrganization()}
           </div>
         </div>
       )}
